Added task_3

+++ Lesson_6/NG_2020_task_3/calculator.cpp
@@ -65,10 +65,10 @@ void Calculator::output()
     {
         switch(action)
         {
-            case '-': number -= ui->Res_bar->text().toInt(); break;
-            case '+': number += ui->Res_bar->text().toInt(); break;
-            case '*': number *= ui->Res_bar->text().toInt(); break;
-            case '/': number /= ui->Res_bar->text().toInt(); break;
+            case '-': number -= ui->Res_bar->text().toFloat(); break;
+            case '+': number += ui->Res_bar->text().toFloat(); break;
+            case '*': number *= ui->Res_bar->text().toFloat(); break;
+            case '/': number /= ui->Res_bar->text().toFloat(); break;
         }
         ui->Res_bar->setText(QString::number(number));
     }
